fixed header, moved insertion form component
