+++ public/index.html
@@ -30,7 +30,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>Geckspance</title>
+    <title>Geckspence</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
